Remove AuthScreen. Use old school LoginScreen instead
